Added versioning and redirection after deleting attachment revision

+++ CM2301/learn/views/attachment.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from django.http import HttpResponse, StreamingHttpResponse
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_http_methods
@@ -15,8 +15,11 @@ def attachment(request, attachment_id):
 @login_required
 def revision_delete(request, revision_id):
     revision = Revision.objects.get(pk=revision_id)
+    if len(revision.attachment.revision_set.all()) < 2:
+        return redirect(revision.attachment.get_absolute_url())
+    
     revision.delete()
-    return HttpResponse('Deleted')
+    return redirect(revision.attachment.get_absolute_url())
     
 def revision(request, revision_id):
     revision = Revision.objects.get(pk=revision_id)
